Апдейт с учетом комментариев
