--- Dodo parking/TicketPrinter.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Dodo_parking
-{
-    public static class TicketPrinter
-    {
-        public static  string Print() 
-        { 
-        return Guid.NewGuid().ToString();
-        }
-
-
-
-    }
-}
